requirements.txt
PyPDF2==3.0.1
requests==2.28.2
python-dotenv==0.21.0
google-genai>=0.1.0
tqdm==4.66.1
PyMuPDF>=1.24.0
Pillow>=10.1.0
openai
anthropic
tenacity>=8.2.0